server/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
SQLAlchemy==2.0.36
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10  # Updated version
marshmallow==3.23.0
marshmallow-sqlalchemy==1.1.0
redis==5.1.1
